refactor: move all components to `components` folder Co-authored-by: Chris Bonifacio <christopher.bonifacio@gmail.com>
